+++ java/ankiweb/src/main/java/com/ankiapp/repository/ResolucaoRepository.java
@@ -3,5 +3,5 @@ package com.ankiapp.repository;
 import com.ankiapp.model.Resolucao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResolucaoRepository extends JpaRepository<Resolucao, Integer> {
+public interface ResolucaoRepository extends JpaRepository<Resolucao, Long> {
 }
